requirements.txt
flask
flask-sqlalchemy
psycopg2-binary
alembic
python-dotenv